Use relative rotations for cross sections
The original method rotated the normals of cross sections from (0, 1, 0)
to the direction of the stem. This would cause twisting under certain
conditions and was fixed with relative rotations.

Each rotation is a rotation from the previous direction to the new direction.
Furthermore, each rotation is concatenated with all previous rotations.