Add an object pool for stems
The intention of this change is that stems can be deallocated and reallocated
without invalidating selection objects. This will enable the generator to
repeatedly generate many stems without having the history hold on to memory.
The leaf selections will still become invalid after the deletion of stems
because new IDs will be generated.