Integrate the generator into the GUI
The generator runs in a separate thread to prevent the GUI from freezing. To
avoid creating additional commands, the generator clears the history and
resets the scene. Upon completion, the grid is replaced with the final octree.

The primitive reset value is increased and additional key commands are added
to change the camera position.